+++ packages/main/src/Calendar.ts
@@ -193,7 +193,7 @@ type SpecialCalendarDateT = {
  * @param {Array<number>} dates The selected dates as UTC timestamps
  * @public
  */
-@event("selected-dates-change", {
+@event<CalendarSelectedDatesChangeEventDetail>("selected-dates-change", {
 	detail: {
 		/**
 		 * @public
@@ -203,6 +203,8 @@ type SpecialCalendarDateT = {
 		 * @public
 		 */
 		values: { type: Array },
+
+		timestamp: { type: Number },
 	},
 })